--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -21,7 +21,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-go@v1
         with:
-          go-version: 1.13.4
+          go-version: 1.15.3
       - run: |
           go build -race  ./...
   test:
@@ -31,7 +31,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-go@v1
         with:
-          go-version: 1.13.4
+          go-version: 1.15.3
       - name: Run tests
         run: |
           go test -race -short ./...
@@ -55,7 +55,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-go@v1
         with:
-          go-version: 1.13.4
+          go-version: 1.15.3
       - run: go mod tidy
       - name: Check for changes in go.mod or go.sum
         run: |