--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module github.com/edwarnicke/serialize
 
 go 1.13
+
+require go.uber.org/goleak v1.1.10