Previously, due to finite channel size it was possible to deadlock
(see new TestNestedAsyncDeadlock test).

Now, we store incoming jobs in a backlog, so that we can accomodate
such cases.