@edwarnicke 2020-10-15
# Before
```
goos: darwin
goarch: amd64
pkg: github.com/edwarnicke/serialize
BenchmarkExecutorAsync
BenchmarkExecutorAsync-16    	 7213419	       163 ns/op
BenchmarkExecutorSync
BenchmarkExecutorSync-16     	 2312421	       528 ns/op
```

#After
```
goos: darwin
goarch: amd64
pkg: github.com/edwarnicke/serialize
BenchmarkExecutorAsync
BenchmarkExecutorAsync-16    	 6077265	       198 ns/op
BenchmarkExecutorSync
BenchmarkExecutorSync-16     	 1779312	       668 ns/op
```